lFormula for gANOVA correlation structure

Construction of the matrices from a formula object, modification of the lme4 function link{lFormula}.

gANOVA_lFormula(
  formula,
  data = NULL,
  REML = TRUE,
  subset,
  weights,
  na.action,
  offset,
  contrasts = NULL,
  control = lmerControl(),
  ...
)

Arguments

formula
a lme4 formula. A formula of type (1|id:g) where id is the grouping variable and g is a factor will have the covariance structure will estimate the same variance for all orthonormal (contr.poly) contrasts in g. WARNINGS The identifier of the grouping variable id must be written as the first terms to the right of the "|" because all other terms will be "reduced" by an orthonormal contrasts. See linkgANOVA for the reduced notation.

data
a data frame. See lmer for more details.

REML
a logical that indicate which criterion to optimize. See lmer for more details.

subset
an expression to selecte a subset of the data. See lmer for more details.

weights
an optional vector of weights. See lmer for more details.

na.action
a function that handle NA's. See lmer for more details.

offset
specify a priori component in the predictor. See lmer for more details.

contrasts
a list of contrasts. See lmer for more details.

control
Some parameters. See lmerControl or lmer for more details.

...
See lFormula or lmer for more details.

See also